+++ src/apis/movie/useGetMovieList.ts
@@ -6,24 +6,7 @@ import {
 } from '../../types/MovieList';
 import { DataQueryKeys } from '../data-query-keys';
 import { endPoints } from '../endPoints';
-
-const snakeCaseToCamelCase = (str: string) => {
-  return str.replace(/([-_][a-z])/gi, $1 => {
-    return $1.toUpperCase().replace('-', '').replace('_', '');
-  });
-};
-
-const convertKeysToCamelCase = (obj: any): any => {
-  if (Array.isArray(obj)) {
-    return obj.map(v => convertKeysToCamelCase(v));
-  } else if (obj !== null && typeof obj === 'object') {
-    return Object.keys(obj).reduce((acc: Record<string, any>, key) => {
-      acc[snakeCaseToCamelCase(key)] = convertKeysToCamelCase(obj[key]);
-      return acc;
-    }, {});
-  }
-  return obj;
-};
+import { convertKeysToCamelCase } from '../../components/ConvertKeysToCamelCase';
 
 export const useGetMovieList = () => {
   return useQuery<NowPlayingMovies>({
