+++ src/components/ConvertKeysToCamelCase/index.tsx
@@ -1,0 +1,17 @@
+const snakeCaseToCamelCase = (str: string) => {
+  return str.replace(/([-_][a-z])/gi, $1 => {
+    return $1.toUpperCase().replace('-', '').replace('_', '');
+  });
+};
+
+export const convertKeysToCamelCase = (obj: any): any => {
+  if (Array.isArray(obj)) {
+    return obj.map(v => convertKeysToCamelCase(v));
+  } else if (obj !== null && typeof obj === 'object') {
+    return Object.keys(obj).reduce((acc: Record<string, any>, key) => {
+      acc[snakeCaseToCamelCase(key)] = convertKeysToCamelCase(obj[key]);
+      return acc;
+    }, {});
+  }
+  return obj;
+};
