+++ src/components/MovieLists/styles.css
@@ -40,3 +40,15 @@
 .now-playing-movie-list-container {
   scrollbar-width: none;
 }
+
+.loader-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  z-index: 9999;
+}
